feat: 자유게시글 작성 api S09P22B210-191

+++ backend/board/src/main/java/com/olock/blockstotck/board/domain/freeboard/persistence/FreePostCommentRepository.java
@@ -3,5 +3,5 @@ package com.olock.blockstotck.board.domain.freeboard.persistence;
 import com.olock.blockstotck.board.domain.freeboard.persistence.entity.FreePostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FreePostCommentRepository extends JpaRepository<FreePostComment, Integer> {
+public interface FreePostCommentRepository extends JpaRepository<FreePostComment, Long> {
 }
