+++ backend/board/src/main/java/com/olock/blockstotck/board/domain/freeboard/persistence/FreePostLikeRepository.java
@@ -3,5 +3,5 @@ package com.olock.blockstotck.board.domain.freeboard.persistence;
 import com.olock.blockstotck.board.domain.freeboard.persistence.entity.FreePostLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FreePostLikeRepository extends JpaRepository<FreePostLike, Integer> {
+public interface FreePostLikeRepository extends JpaRepository<FreePostLike, Long> {
 }
